requirements.txt
numpy~=2.1.3
torch~=2.5.1
transformers~=4.47.0
networkx~=3.4.2
gym~=0.26.2
ray~=2.40.0
plotly~=5.24.1
matplotlib~=3.9.3
seaborn~=0.13.2
pandas~=2.2.3
pip~=23.2.1
attrs~=24.2.0
wheel~=0.41.2
protobuf~=5.29.1
pillow~=11.0.0
Jinja2~=3.1.4
setuptools~=68.2.0
packaging~=24.2
pyparsing~=3.2.0
typing_extensions~=4.12.2
PyYAML~=6.0.2
lz4
dm-tree
pyarrow~=18.1.0
scipy~=1.14.1